Aym 93 (#42)
* bambumka

* calis artık

---------

+++ review/internal/src/main/java/gtu/cse/se/altefdirt/aymoose/review/internal/infra/adapter/rest/dto/ReviewResponseDTO.java
@@ -17,7 +17,7 @@ public record ReviewResponseDTO(
     public static ReviewResponseDTO fromView(ReviewView reviewView) {
         return new ReviewResponseDTO(reviewView.id(),
                 reviewView.author(),
-                reviewView.rating(),
+                reviewView.rating().shortValue(),
                 reviewView.title(),
                 reviewView.content(),
                 reviewView.createdAt().toString(),
